+++ packages/plugins/Scale/src/index.ts
@@ -1,10 +1,10 @@
 import Vue from 'vue'
-import { PluginOptions } from '@polar/lib-custom-types'
+import { ScaleConfiguration } from '@polar/lib-custom-types'
 import { Scale } from './components'
 import language from './language'
 import { makeStoreModule } from './store'
 
-export default (options: PluginOptions) => (instance: Vue) =>
+export default (options: ScaleConfiguration) => (instance: Vue) =>
   instance.$store.dispatch('addComponent', {
     name: 'scale',
     plugin: Scale,